requirements.txt
bs4
requests
openai
langchain
langchain_community
python-dotenv
sentence-transformers
chromadb
streamlit
langgraph 
nltk
rouge_score